MAS - Ajout du module NN_Pytorch_Lambda pour algo_lambda
